Restructure AWS driver packages under dynamodb/ and s3/ subdirectories
