+++ driver/aws/dynamodb/dynamojournal/doc.go
