+++ driver/aws/dynamodb/dynamojournal/journal.go
