+++ driver/aws/dynamodb/dynamojournal/schema.go
