+++ driver/aws/dynamodb/dynamojournal/store.go
