+++ driver/aws/s3/s3journal/compact.go
