+++ driver/aws/s3/s3kv/store_test.go
@@ -1,18 +1,18 @@
-package s3set_test
+package s3kv_test
 
 import (
 	"testing"
 
 	"github.com/aws/aws-sdk-go-v2/service/s3"
 	"github.com/dogmatiq/persistencekit/driver/aws/internal/s3x"
-	. "github.com/dogmatiq/persistencekit/driver/aws/s3set"
+	. "github.com/dogmatiq/persistencekit/driver/aws/s3/s3kv"
 	"github.com/dogmatiq/persistencekit/internal/x/xtesting"
-	"github.com/dogmatiq/persistencekit/set"
+	"github.com/dogmatiq/persistencekit/kv"
 )
 
 func TestStore(t *testing.T) {
 	client, bucket := setup(t)
-	set.RunTests(
+	kv.RunTests(
 		t,
 		NewBinaryStore(client, bucket),
 	)
@@ -20,7 +20,7 @@ func TestStore(t *testing.T) {
 
 func BenchmarkStore(b *testing.B) {
 	client, bucket := setup(b)
-	set.RunBenchmarks(
+	kv.RunBenchmarks(
 		b,
 		NewBinaryStore(client, bucket),
 	)
